Resolves #193
