+++ assembler/src.txt
@@ -7,6 +7,7 @@ src/data/project.cpp
 src/details/shader_cache.cpp
 src/details/spirv.cpp
 
+src/generators/meta.cpp
 src/generators/models.cpp
 src/generators/project.cpp
 src/generators/shader.cpp
