+++ assembler/src/importers/meta.cpp
@@ -5,13 +5,14 @@
 #include <psl/meta.hpp>
 #include <psl/serialization/serializer.hpp>
 
+#if !defined(PE_GLES)
+	#include "GLES3/gl32.h"
+#endif
+
+#include "core/gles/conversion.hpp"
 #include "details/texture_utils.hpp"
 #include <core/meta/texture.hpp>
 #include <psl/terminal_utils.hpp>
-#if defined(PE_GLES)
-	#include "core/gles/conversion.hpp"
-#endif
-#include <stdexcept>
 
 namespace assembler::importer {
 void generate_texture_meta(std::filesystem::path path, core::meta::texture_t* texture_meta) {
@@ -22,10 +23,6 @@ void generate_texture_meta(std::filesystem::path path, core::meta::texture_t* te
 	auto view = psl::array_view<std::byte>((std::byte*)data.data(), data.size());
 
 	if(utility::ktx::is_ktx(view)) {
-#if !defined(PE_GLES)
-		// todo(jdl): we can generate the gl headers by default. We only really need the define values.
-		throw std::runtime_error("ktx files are only supported when the GLES backend is enabled.");
-#else
 		auto header = utility::ktx::decode(psl::array_view<std::byte>((std::byte*)data.data(), data.size()));
 		texture_meta->width(header.pixelWidth);
 		texture_meta->height(header.pixelHeight);
@@ -33,7 +30,6 @@ void generate_texture_meta(std::filesystem::path path, core::meta::texture_t* te
 		texture_meta->mip_levels(header.numberOfMipmapLevels);
 		texture_meta->format(core::gfx::conversion::to_format(header.glInternalFormat, header.glFormat, header.glType));
 		psl_assert(texture_meta->format() != core::gfx::format_t::undefined);
-#endif
 	} else if(utility::dds::is_dds(view)) {
 		auto header = utility::dds::decode(psl::array_view<std::byte>((std::byte*)data.data(), data.size()));
 		texture_meta->width(header.dwWidth);
